requirements.txt
asgiref==3.5.0
Django==4.0.3
django-ckeditor==6.2.0
django-js-asset==2.0.0
djangorestframework==3.13.1
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
psycopg2-binary==2.9.3
